pyproject.toml
[project]
name = "Financial-News-Intelligent-System---Enterprise-Level"
version = "0.1.0"
description = "This is a financial news intelligent system made for finance analysts to reduce their workload in analyzing news and making an impactful decision."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.8",
    "feedparser>=6.0.12",
    "finnhub-python>=2.4.27",
    "loguru>=0.7.3",
    "newsapi-python>=0.2.7",
    "openai>=2.33.0",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "sqlalchemy>=2.0.48",
]
